+++ tests/tools/decompiler_tests.py
@@ -400,6 +400,14 @@ class ParseArgsTests(ToolTestsBase):
         args = parse_args(['decompiler.py', '--optimizations', 'none', 'prog.exe'])
         self.assertEqual(args.decomp_optimizations, 'none')
 
+    def test_decomp_unreach_funcs_is_parsed_correctly_short_form(self):
+        args = parse_args(['decompiler.py', '-K', 'prog.exe'])
+        self.assertTrue(args.decomp_unreach_funcs)
+
+    def test_decomp_unreach_funcs_is_parsed_correctly_long_form(self):
+        args = parse_args(['decompiler.py', '--keep-unreach-funcs', 'prog.exe'])
+        self.assertTrue(args.decomp_unreach_funcs)
+
     def test_generate_archive_is_set_to_false_when_with_archive_not_given(self):
         args = parse_args(['decompiler.py', 'prog.exe'])
         self.assertFalse(args.generate_archive)
@@ -652,6 +660,15 @@ class MainTests(ToolTestsBase):
 
         self.assert_decompilation_was_started_also_with(
             decomp_optimizations='none'
+        )
+
+    def test_sets_decomp_unreach_funcs_when_given(self):
+        self.call_main_with_standard_arguments_and(
+            '--keep-unreach-funcs'
+        )
+
+        self.assert_decompilation_was_started_also_with(
+            decomp_unreach_funcs=True
         )
 
     def test_generates_and_saves_archive_when_requested(self):