+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ dev
 * Added support for selecting a different style for naming of variables.
 * Added support for selecting the type of optimizations to be performed by the
   decompiler.
+* Added support for decompilation of unreachable functions.
 
 0.3 (2016-07-17)
 ----------------